+++ src/components/PdfHighlighter.tsx
@@ -7,6 +7,7 @@ import {
   NullL10n,
   PDFLinkService,
   PDFViewer,
+  PDFFindController,
 } from "pdfjs-dist/legacy/web/pdf_viewer";
 import type {
   IHighlight,
@@ -71,8 +72,11 @@ interface Props<T_HT> {
   highlights: Array<T_HT>;
   onScrollChange: () => void;
   scrollRef: (scrollTo: (highlight: T_HT) => void) => void;
+  findRefs: (findPrev: () => void, findNext: () => void) => void;
   pdfDocument: PDFDocumentProxy;
   pdfScaleValue: string;
+  searchValue: string;
+  onSearch: (currentMatch: number, totalMatchCount: number) => void;
   onSelectionFinished: (
     position: ScaledPosition,
     content: { text?: string; image?: string },
@@ -90,6 +94,8 @@ export class PdfHighlighter<T_HT extends IHighlight> extends PureComponent<
 > {
   static defaultProps = {
     pdfScaleValue: "auto",
+    searchValue: "",
+    onSearch: () => { },
   };
 
   state: State<T_HT> = {
@@ -140,6 +146,7 @@ export class PdfHighlighter<T_HT extends IHighlight> extends PureComponent<
       const { ownerDocument: doc } = ref;
       eventBus.on("textlayerrendered", this.onTextLayerRendered);
       eventBus.on("pagesinit", this.onDocumentReady);
+      eventBus.on("updatetextlayermatches", this.calculateMatchProgress);
       doc.addEventListener("selectionchange", this.onSelectionChange);
       doc.addEventListener("keydown", this.handleKeyDown);
       doc.defaultView?.addEventListener("resize", this.debouncedScaleValue);
@@ -148,6 +155,7 @@ export class PdfHighlighter<T_HT extends IHighlight> extends PureComponent<
       this.unsubscribe = () => {
         eventBus.off("pagesinit", this.onDocumentReady);
         eventBus.off("textlayerrendered", this.onTextLayerRendered);
+        eventBus.off("updatetextlayermatches", this.calculateMatchProgress);
         doc.removeEventListener("selectionchange", this.onSelectionChange);
         doc.removeEventListener("keydown", this.handleKeyDown);
         doc.defaultView?.removeEventListener(
@@ -170,6 +178,13 @@ export class PdfHighlighter<T_HT extends IHighlight> extends PureComponent<
     if (prevProps.pdfScaleValue != this.props.pdfScaleValue) {
       this.viewer.currentScaleValue = this.props.pdfScaleValue;
     }
+    if (prevProps.searchValue != this.props.searchValue) {
+      this.eventBus.dispatch("find", {
+        query: this.props.searchValue,
+        highlightAll: true,
+        phraseSearch: true,
+      });
+    }
   }
 
   init() {
@@ -185,6 +200,10 @@ export class PdfHighlighter<T_HT extends IHighlight> extends PureComponent<
         textLayerMode: 2,
         removePageBorders: true,
         linkService: this.linkService,
+        findController: new PDFFindController({
+          eventBus: this.eventBus,
+          linkService: this.linkService,
+        }),
         l10n: NullL10n,
       });
 
@@ -194,6 +213,48 @@ export class PdfHighlighter<T_HT extends IHighlight> extends PureComponent<
     // debug
     (window as any).PdfViewer = this;
   }
+
+  calculateMatchProgress = () => {
+    let totalMatchCount = 0;
+    let currentMatchIndex = 0;
+    let hasCurrentMatchIndex = false;
+
+    this.viewer.findController?.pageMatches?.forEach(
+      (pageMatches: number[], pageIndex: number) => {
+        if (
+          !hasCurrentMatchIndex &&
+          pageIndex === this.viewer.findController?.selected?.pageIdx
+        ) {
+          currentMatchIndex =
+            totalMatchCount + this.viewer.findController?.selected?.matchIdx;
+          hasCurrentMatchIndex = true;
+        }
+        totalMatchCount += pageMatches.length;
+      }
+    );
+
+    const currentMatch = totalMatchCount === 0 ? 0 : currentMatchIndex + 1;
+    this.props.onSearch(currentMatch, totalMatchCount);
+  };
+
+  goToNextMatch = () => {
+    const { searchValue } = this.props;
+    this.eventBus.dispatch("find", {
+      query: searchValue,
+      highlightAll: true,
+      type: "again",
+    });
+  };
+
+  goToPreviousMatch = () => {
+    const { searchValue } = this.props;
+    this.eventBus.dispatch("find", {
+      query: searchValue,
+      highlightAll: true,
+      findPrevious: true,
+      type: "again",
+    });
+  };
 
   componentWillUnmount() {
     this.unsubscribe();
@@ -411,11 +472,12 @@ export class PdfHighlighter<T_HT extends IHighlight> extends PureComponent<
   };
 
   onDocumentReady = () => {
-    const { scrollRef } = this.props;
+    const { scrollRef, findRefs } = this.props;
 
     this.handleScaleValue();
 
     scrollRef(this.scrollTo);
+    findRefs(this.goToPreviousMatch, this.goToNextMatch);
   };
 
   onSelectionChange = () => {